pyproject.toml
[project]
name = "simple-test-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyinstaller>=6.15.0",
    "pyodbc>=5.2.0",
    "pyside6>=6.9.2",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.11",
    "sqlalchemy>=2.0.43",
    "werkzeug>=3.1.3",
]

[tool.ruff]
# Target Python version
target-version = "py311"

# Maximum line length (affects lint + formatter)
line-length = 88

# Enable these rule families:
# - E: pycodestyle (PEP8 errors)
# - F: pyflakes (unused imports, undefined names, etc.)
# - I: isort (import sorting)
# - B: flake8-bugbear (common pitfalls)
# - UP: pyupgrade (use modern syntax)
select = ["E", "F", "I", "B", "UP"]

# Rules you *don’t* want (common ignores)
ignore = [
    "E501",  # line too long (handled by formatter instead)
    "B905",  # zip() without strict=... (can be noisy in legacy code)
]

# Automatically fixable rules (what `ruff --fix` will apply)
fixable = ["ALL"]

# Exclude certain directories
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "build",
    "dist",
    "migrations"
]

# Per-file ignores (for special cases)
[tool.ruff.per-file-ignores]
"tests/*" = ["D", "S101"]  # ignore missing docstrings & asserts in tests

# Optional: formatter config (if you use `ruff format`)
[tool.ruff.format]
quote-style = "double"   # or "single"
indent-style = "space"   # or "tab"
skip-magic-trailing-comma = false
line-ending = "lf"
